deleted unit test file, it is no longer needed updates to unit test files for code coverage
