+++ jmaqs-webservices-jdk11/src/test/java/com/magenic/jmaqs/webservices/jdk11/WebServiceDriverPatchUnitTest.java
@@ -23,7 +23,15 @@ public class WebServiceDriverPatchUnitTest extends BaseWebServiceTest {
    */
   private static final String baseUrl = WebServiceConfig.getWebServiceUri();
 
+  /**
+   * Product object to be used in a test.
+   */
   private final Product product = new Product(4, "ff", "ff", BigDecimal.valueOf(3.25));
+
+  /**
+   * The web service driver to be used in a test.
+   */
+  private final WebServiceDriver webServiceDriver = new WebServiceDriver(HttpClientFactory.getDefaultClient());
 
   /**
    * Verify the string status code.
@@ -192,5 +200,71 @@ public class WebServiceDriverPatchUnitTest extends BaseWebServiceTest {
         MediaType.PLAIN_TEXT, "", false);
     String expected = "{\"Message\":\"The requested resource does not support http method 'PATCH'.\"}";
     Assert.assertEquals(result.body(), expected);
+  }
+
+
+  /**
+   * Test type parameterized Patch request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void patchTypeParamWithExpectedStatus() throws IOException, InterruptedException {
+    var req = WebServiceUtilities.createStringEntity(product, MediaType.APP_JSON);
+    var res = this.webServiceDriver.patch(baseUrl +"/api/XML_JSON/Patch/1",
+        MediaType.APP_JSON, req, Product.class, HttpStatus.OK);
+    Assert.assertNotNull(res);
+  }
+
+  /**
+   * Test Patch request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void patchWithExpectedStatus() throws IOException, InterruptedException {
+    var req = WebServiceUtilities.createStringEntity(product, MediaType.APP_JSON);
+    var res = this.webServiceDriver.patch(
+        baseUrl+ "/api/XML_JSON/Patch/1", MediaType.APP_JSON, req, HttpStatus.OK);
+    Assert.assertNotNull(res);
+  }
+
+  /**
+   * Test more params Patch request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void patchMoreParamsWithExpectedStatus() throws IOException, InterruptedException {
+    var req = WebServiceUtilities.createStringEntity(product, MediaType.APP_JSON);
+    var res = this.webServiceDriver.patch(baseUrl + "/api/XML_JSON/Patch/1",
+        MediaType.APP_JSON, req, MediaType.APP_JSON, HttpStatus.OK);
+    Assert.assertNotNull(res);
+  }
+
+  /**
+   * Test more params Patch with response request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void patchMoreParamsWithResponseWithExpectedStatus()
+      throws IOException, InterruptedException {
+    var res = this.webServiceDriver.patch(baseUrl + "/api/XML_JSON/Patch/1",
+        MediaType.APP_JSON, WebServiceUtilities.serializeJson(product), MediaType.APP_JSON, HttpStatus.OK);
+    Assert.assertNotNull(res);
+  }
+
+  /**
+   * Test Patch with response request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void patchWithResponseWithExpectedStatus() throws IOException, InterruptedException {
+    var req = WebServiceUtilities.createStringEntity(product, MediaType.APP_JSON);
+    var res = this.webServiceDriver.patch(
+        baseUrl + "/api/XML_JSON/Patch/1", MediaType.APP_JSON, req, HttpStatus.OK);
+    Assert.assertNotNull(res);
   }
 }
